Tasks: fixed default list not selected by default in task dialog

+++ www/modules/email/model/ImapMailbox.php
@@ -268,11 +268,14 @@ class ImapMailbox extends \GO\Base\Model {
 		}
 		$imap = $this->getAccount()->openImapConnection($this->name);
 		$success = true;
-		foreach ($imap->get_folders($this->_account->trash) as $folder) {
-			if($folder['name'] == $this->_account->trash || empty($folder['name'])) {
-				continue;
+
+		if($this->_account->trash && $this->_account->trash == $this->name) {
+			foreach ($imap->get_folders($this->_account->trash) as $folder) {
+				if ($folder['name'] == $this->_account->trash || empty($folder['name'])) {
+					continue;
+				}
+				$success = $success && $imap->delete_folder($folder['name']);
 			}
-			$success = $success &&$imap->delete_folder($folder['name']);
 		}
 		$sort = $imap->sort_mailbox();
 		return $imap->delete($sort);
